login/register funcionando

+++ backend/main/resources/Producto.py
@@ -2,19 +2,30 @@ from flask_restful import Resource
 from flask import request, jsonify
 from .. import db
 from main.models import ProductoModel, UsuarioModel
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from main.auth.decorators import admin_required
 
 
 class Producto(Resource):
+    @jwt_required()
     def get(self, id):
         producto = db.session.query(ProductoModel).get_or_404(id)
-        return producto.to_json()
+        if UsuarioModel.role == 'proveedor' or UsuarioModel.role == 'admin':
+            return producto.to_json()
+        else:
+            return '', 404
 
+    @jwt_required()
     def delete(self, id):
         producto = db.session.query(ProductoModel).get_or_404(id)
-        db.session.delete(producto)
-        db.session.commit()
-        return '', 204
+        if UsuarioModel.role == 'proveedor' or UsuarioModel.role == 'admin':
+            db.session.delete(producto)
+            db.session.commit()
+            return '', 204
+        else:
+            return '', 404
 
+    @jwt_required()
     def put(self, id):
         producto = db.session.query(ProductoModel).get_or_404(id)
         data = request.get_json().items()
@@ -26,6 +37,7 @@ class Producto(Resource):
 
 
 class Productos(Resource):
+    @jwt_required()
     def get(self):
         page = 1
         per_page = 10
@@ -44,6 +56,7 @@ class Productos(Resource):
                         'page': page
                         })
 
+    @jwt_required()
     def post(self):
         producto = ProductoModel.from_json(request.get_json())
         proveedor = db.session.query(UsuarioModel).get_or_404(producto.usuarioid)
